+++ src/DessserBackEndOCaml.ml
@@ -1717,7 +1717,8 @@ end
 
 include DessserBackEndCLike.Make (Config)
 
-let compile ?dev_mode ?extra_search_paths ?optim ~link ?dst_fname ?comment ?outro compunit =
+let compile ?dev_mode ?extra_search_paths ?optim ~link ?dst_fname ?comment
+            ?outro ?(keep_temp_files=false) compunit =
   let dst_fname =
     (match dst_fname with
     | None ->
@@ -1734,7 +1735,7 @@ let compile ?dev_mode ?extra_search_paths ?optim ~link ?dst_fname ?comment ?outr
     print_definitions oc compunit ;
     Option.may (String.print oc) outro) ;
   run_cmd cmd ;
-  ignore_exceptions Unix.unlink src_fname ;
+  if not keep_temp_files then ignore_exceptions Unix.unlink src_fname ;
   dst_fname
 
 (* Compile and dynload the given compunit.
